03-DB_TELEGRAM/requirements.txt
azure-cognitiveservices-speech==1.43.0
google-genai>=1.0.0
elevenlabs>=1.0.0
python-dotenv==1.0.1
arabic-reshaper==3.0.0
python-bidi==0.4.2
chromadb>=0.6.0

